Simulate the Monty Hall Problem
Bayes' Theorem Example: Monty Hall Problem
A very famous stats mind trick... You might have seen this already.

The Monty Hall Problem was originally posed (and solved) in a letter by Steve Selvin to the American Statistician in 1975. It became famous as a question from reader Craig F. Whitaker's letter quoted in Marilyn vos Savant's "Ask Marilyn" column in Parade magazine in 1990

You are playing a TV game show and are shown 2 doors. One has a car behind it, the other a goat. What are your chances of picking the door with the car?

OK, now there are 3 doors: one with a car, two with goats. The game show host asks you to pick a door, but not to open it yet. Then the host opens one of the other two doors (that you did not pick) and that has a goat. The host offers you the opportunity to switch doors.

One player decides to switch
Another player prefers to stay with the previous choice
https://upload.wikimedia.org/wikipedia/commons/thumb/3/3f/Monty_open_door.svg/180px-Monty_open_door.svg.png

Now a third player comes in who has not seen what happened earlier. They pick a door.
Which of the three players is most likely to win?

You might know or remember the answer already... but don't think now! Let's simulate it.

Simulate the Monty Hall Problem
Simulate three doors, one car, and two goats.

Simulate three players: the switcher, the conservative, and the newcomer.

Record who wins.

Repeat it many times.

Which player do you want to be?

What would happen if you had 100 doors to choose from and the presenter opens 98 or them?