Exoplanets
Exoplanets are planets outside of the solar system. The existence of exoplanets was first confirmed in 1992. Since then the number of detected exoplanets exploded, and right now there are more than 5000! Some as are light as our Moon, and some are as heavy as 30 times the mass of Jupyter!

NASA hosts a public database with all the exoplanets discovered:

https://exoplanetarchive.ipac.caltech.edu/cgi-bin/TblView/nph-tblView?app=ExoTbls&config=PS&constraint=default_flag%20%3E0

These planets orbit far away stars. This plots shows the 2d distributions of the size of the orbit (x axis) and the mass of the planet (y axis).